+++ models/event/events_suite_test.go
@@ -1,0 +1,35 @@
+package event
+
+import (
+	"context"
+	"testing"
+
+	"github.com/artie-labs/transfer/models"
+
+	"github.com/artie-labs/transfer/lib/config"
+
+	"github.com/artie-labs/transfer/lib/mocks"
+	"github.com/stretchr/testify/suite"
+)
+
+type EventsTestSuite struct {
+	suite.Suite
+	fakeStore *mocks.FakeStore
+	ctx       context.Context
+}
+
+func (e *EventsTestSuite) SetupTest() {
+	e.ctx = context.Background()
+	e.ctx = config.InjectSettingsIntoContext(e.ctx, &config.Settings{
+		Config: &config.Config{
+			FlushIntervalSeconds: 10,
+			BufferRows:           10,
+		},
+	})
+
+	e.ctx = models.LoadMemoryDB(e.ctx)
+}
+
+func TestEventsTestSuite(t *testing.T) {
+	suite.Run(t, new(EventsTestSuite))
+}
